+++ CHANGELOG.md
@@ -6,6 +6,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Checkpoint] - 2026-02-03 08:37:02
+
+### Changed Files
+- `src/contexts/AuthContext.tsx`
+- `src/contexts/ImoContext.tsx`
+- `src/services/agency/AgencyService.ts`
+- `src/services/imo/ImoService.ts`
+
+### Statistics
+```
+ 4 files changed, 71 insertions(+), 60 deletions(-)
+```
+
 ## [Checkpoint] - 2026-02-02 19:37:31
 
 ### Changed Files